# docs/openapi-legajo.yaml
openapi: 3.0.3
info:
  title: Legajo API
  description: >-
    REST contract for Legajo's six similarity capabilities and its
    hierarchical clustering with four linkage criteria over the reference
    corpus. This file is the contract's source of
    truth: it describes exactly what the running server accepts and
    returns today, never an aspiration. The frontend generates its TypeScript
    types from this file with `npm run api:types`, and CI fails on drift.
  version: "1.0.0"
  contact:
    name: Legajo backend
    url: https://github.com/diegnghtmr/legajo
  license:
    name: UNLICENSED
servers:
  - url: /
    description: Same-origin (Render in production; localhost in development)

# No authentication scheme exists anywhere in this backend (none is defined for
# this demo API); declared explicitly, rather than left implicit, so linters and readers see
# an intentional "no security" rather than an oversight.
security: []
...
paths:
...
  /api/v1/similarity/algorithms:
    get:
      operationId: listSimilarityAlgorithms
      tags: [similarity]
      summary: The catalogue of the six similarity capabilities
      description: Exactly six ids, including `needleman-wunsch`.
      responses:
        "200":
          description: The six capabilities, in registration order.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AlgorithmSummary"
        default:
          $ref: "#/components/responses/UnexpectedError"
  /api/v1/similarity/compare:
    post:
      operationId: compareSimilarity
      tags: [similarity]
      summary: Multi-algorithm pairwise similarity for two documents
      description: >-
        `algorithmIds` is optional; omitted, `null` or empty defaults to all
        six capabilities, so the default request already exercises every
        capability. A `null` or blank entry inside `algorithmIds` is a
        malformed request.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CompareRequest"
      responses:
        "200":
          description: One result per requested (or defaulted) algorithm.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AlgorithmSimilarity"
        "400":
          description: >-
            An unknown `algorithmIds` entry (`unknown-algorithm`), an unknown
            `documentIdA`/`documentIdB` (`unknown-document`), a blank
            required field or a `null`/blank entry in `algorithmIds`
            (`about:blank`, framework-detected) — 400
            because these values arrive in the request body.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
              examples:
                unknownAlgorithm:
                  value:
                    type: "urn:legajo:problem:unknown-algorithm"
                    title: "Bad Request"
                    status: 400
                    detail: "no similarity algorithm registered with id: bogus"
                unknownDocument:
                  value:
                    type: "urn:legajo:problem:unknown-document"
                    title: "Bad Request"
                    status: 400
                    detail: "no corpus document with id: does-not-exist"
        "503":
          $ref: "#/components/responses/EmbeddingApiUnavailable"
        default:
          $ref: "#/components/responses/UnexpectedError"
  /api/v1/similarity/matrix:
    post:
      operationId: similarityMatrix
      tags: [similarity]
      summary: m x m similarity matrix for one algorithm over a document selection
      description: >-
        m = selection size, business rule m ∈ [3, n] (3–20 in the reference
        corpus); the response is a plain m x m nested array with no envelope,
        row/column order matching `documentIds` as sent.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MatrixRequest"
      responses:
        "200":
          description: An m x m grid; row i, column j is the pair `(documentIds[i], documentIds[j])`.
          content:
            application/json:
              schema:
                type: array
                items:
                  type: array
                  items:
                    $ref: "#/components/schemas/SimilarityResult"
        "400":
          description: >-
            An unknown `algorithmId` (`unknown-algorithm`), an unknown entry
            in `documentIds` (`unknown-document`), or a selection smaller than
            3 elements or containing duplicates (`invalid-selection`).
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
              examples:
                invalidSelection:
                  value:
                    type: "urn:legajo:problem:invalid-selection"
                    title: "Bad Request"
                    status: 400
                    detail: "documentIds must not be empty"
        "503":
          $ref: "#/components/responses/EmbeddingApiUnavailable"
        default:
          $ref: "#/components/responses/UnexpectedError"
  /api/v1/similarity/{algorithmId}/trace:
    get:
      operationId: similarityTrace
      tags: [similarity]
      summary: The complete audit trace for one capability and one document pair
      description: >-
        Returns the **complete** trace with no truncation parameter.
        The trace pair is identified by two required query
        parameters, since the path alone does not identify a pair for a GET.
      parameters:
        - name: algorithmId
          in: path
          required: true
          description: One of the six similarity capability ids.
          schema:
            $ref: "#/components/schemas/AlgorithmId"
        - name: documentIdA
          in: query
          required: true
          schema:
            type: string
        - name: documentIdB
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: The complete trace; its shape depends on `algorithmId`.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AlgorithmTrace"
        "400":
          description: >-
            An unknown `documentIdA`/`documentIdB` query value
            (`unknown-document`) — 400 because these are
            query parameters, not the path segment.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "404":
          description: >-
            An unknown `algorithmId` path segment (`unknown-algorithm`) — 404
            because this value is a path segment.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
              example:
                type: "urn:legajo:problem:unknown-algorithm"
                title: "Not Found"
                status: 404
                detail: "no similarity algorithm registered with id: bogus"
        "503":
          $ref: "#/components/responses/EmbeddingApiUnavailable"
        default:
          $ref: "#/components/responses/UnexpectedError"
...
components:
...
  responses:
    EmbeddingApiUnavailable:
      description: >-
        A network failure, timeout, rejection, missing API key, or dimension
        mismatch from the live `embedding-api` provider. Applies to every
        endpoint that can use the `embedding-api` capability/representation.
        In live mode, the server fetches missing vectors from the configured
        OpenAI-compatible embeddings endpoint at request time; any failure of
        that call is mapped to this response by
        `ProblemDetailExceptionHandler`.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetail"
          example:
            type: "urn:legajo:problem:embedding-api-unavailable"
            title: "Service Unavailable"
            status: 503
            detail: "The live embedding API is currently unavailable."
    UnexpectedError:
      description: >-
        Any status this operation's explicit responses do not name: a
        framework-detected client mistake (malformed JSON, missing
        parameter, unsupported method/media type, unknown route, a blank
        Bean Validation field — `about:blank`, Spring's own handling) or an
        unexpected server fault (500, generic detail, no exception message
        or class name ever exposed).
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetail"
          example:
            title: "Internal Server Error"
            status: 500
            detail: "An unexpected error occurred."
  schemas:
    AlgorithmId:
      type: string
      description: One of the six fixed similarity capability ids.
      enum:
        - levenshtein
        - needleman-wunsch
        - jaccard
        - tfidf-cosine
        - embedding-local
        - embedding-api
    AlgorithmKind:
      type: string
      enum: [CLASSIC, AI]
...
    ProblemDetail:
      type: object
      description: >-
        RFC 9457 Problem Detail, as Spring Framework 7 serializes it natively.
        `type` carries one of this API's fixed URNs, or is entirely absent
        (never a literal JSON `null`) when the error is one Spring's own
        machinery classifies without a domain-specific cause.
      required: [title, status]
      properties:
        type:
          type: string
          format: uri
          description: >-
            A fixed `urn:legajo:problem:*` URN (never a resolvable URL), or
            omitted entirely for a framework-detected or unexpected error
            (Spring's default is `about:blank`, which it also omits from the
            serialized body).
          enum:
            - "urn:legajo:problem:unknown-document"
            - "urn:legajo:problem:unknown-algorithm"
            - "urn:legajo:problem:unknown-representation"
            - "urn:legajo:problem:unknown-linkage"
            - "urn:legajo:problem:invalid-cut"
            - "urn:legajo:problem:invalid-selection"
            - "urn:legajo:problem:embedding-api-unavailable"
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        instance:
          type: string
          format: uri-reference
          description: >-
            Spring populates this with the request path (a relative
            reference, e.g. `/api/v1/corpus/does-not-exist`), never an
            absolute URI — `uri-reference` (RFC 3986 URI-reference) is the
            correct, narrower format for that shape, unlike `type` above,
            which is always an absolute URN or omitted.
...
    AlgorithmSummary:
      type: object
      required: [id, displayName, kind]
      properties:
        id:
          $ref: "#/components/schemas/AlgorithmId"
        displayName:
          type: string
        kind:
          $ref: "#/components/schemas/AlgorithmKind"
    SimilarityResult:
      type: object
      description: Common result envelope, as returned over the wire (with `cached`).
      required: [normalizedScore, computedNanos, cached, degenerate, stemming]
      properties:
        normalizedScore:
          type: number
          format: double
          minimum: 0
          maximum: 1
        rawValue:
          type: number
          format: double
          nullable: true
          description: "null only in the documented degenerate cases (e.g. the TF-IDF null vector)."
        computedNanos:
          type: integer
          format: int64
        cached:
          type: boolean
        degenerate:
          type: boolean

        stemming:
          type: boolean
          description: >-
            Whether Porter stemming was applied to the token streams behind
            this result (the server's fixed `legajo.preprocess.stemming`
            setting; `false` by default). Embedding capabilities do not
            preprocess text, so for them it only echoes the setting.
    AlgorithmSimilarity:
      type: object
      required: [algorithmId, result]
      properties:
        algorithmId:
          $ref: "#/components/schemas/AlgorithmId"
        result:
          $ref: "#/components/schemas/SimilarityResult"
    CompareRequest:
      type: object
      required: [documentIdA, documentIdB]
      properties:
        documentIdA:
          type: string
        documentIdB:
          type: string
        algorithmIds:
          type: array
          nullable: true
          description: >-
            Omitted, null or empty defaults to all six capabilities. Entries
            must be non-null, non-blank ids.
          items:
            $ref: "#/components/schemas/AlgorithmId"
    MatrixRequest:
      type: object
      required: [algorithmId, documentIds]
      properties:
        algorithmId:
          $ref: "#/components/schemas/AlgorithmId"
        documentIds:
          type: array
          description: "Business rule: size m ∈ [3, n] (3–20 in the reference corpus), no duplicates. Entries must be non-null, non-blank ids."
          items:
            type: string
            minLength: 1
    MatrixCell:
      type: object
      required: [row, col]
      properties:
        row:
          type: integer
        col:
          type: integer
    DpOperationKind:
      type: string
      description: >-
        MATCH/SUBSTITUTION/INSERTION/DELETION classify a Levenshtein step;
        MATCH/MISMATCH/GAP classify a Needleman-Wunsch step.
      enum: [MATCH, SUBSTITUTION, INSERTION, DELETION, MISMATCH, GAP]
    DpTraceStep:
      type: object
      required: [from, to, operation]
      properties:
        from:
          $ref: "#/components/schemas/MatrixCell"
        to:
          $ref: "#/components/schemas/MatrixCell"
        operation:
          $ref: "#/components/schemas/DpOperationKind"
    DpMatrixTrace:
      type: object
      description: Shared by `levenshtein` and `needleman-wunsch`.
      required: [algorithmId, rowLabels, columnLabels, matrix, optimalPath, operations, stemming]
      properties:
        algorithmId:
          $ref: "#/components/schemas/AlgorithmId"
        stemming:
          type: boolean
          description: >-
            Whether Porter stemming was applied to the token streams behind
            this result (the server's fixed `legajo.preprocess.stemming`
            setting; `false` by default). Embedding capabilities do not
            preprocess text, so for them it only echoes the setting.
        rowLabels:
          type: array
          items:
            type: string
        columnLabels:
          type: array
          items:
            type: string
        matrix:
          type: array
          description: The complete DP matrix, never truncated.
          items:
            type: array
            items:
              type: number
              format: double
        optimalPath:
          type: array
          items:
            $ref: "#/components/schemas/MatrixCell"
        operations:
          type: array
          items:
            $ref: "#/components/schemas/DpTraceStep"
    JaccardTrace:
      type: object
      required: [algorithmId, setA, setB, intersectionSize, unionSize, intersection, union, coefficient, stemming]
      properties:
        algorithmId:
          $ref: "#/components/schemas/AlgorithmId"
        stemming:
          type: boolean
          description: >-
            Whether Porter stemming was applied to the token streams behind
            this result (the server's fixed `legajo.preprocess.stemming`
            setting; `false` by default). Embedding capabilities do not
            preprocess text, so for them it only echoes the setting.
        setA:
          type: array
          items:
            type: string
        setB:
          type: array
          items:
            type: string
        intersectionSize:
          type: integer
        unionSize:
          type: integer
        intersection:
          type: array
          items:
            type: string
        union:
          type: array
          items:
            type: string
        coefficient:
          type: number
          format: double
    TfIdfTermTrace:
      type: object
      required:
        - term
        - frequencyA
        - frequencyB
        - documentFrequency
        - tfA
        - tfB
        - idf
        - rawWeightA
        - rawWeightB
        - normalizedWeightA
        - normalizedWeightB
      properties:
        term:
          type: string
        frequencyA:
          type: integer
        frequencyB:
          type: integer
        documentFrequency:
          type: integer
        tfA:
          type: number
          format: double
        tfB:
          type: number
          format: double
        idf:
          type: number
          format: double
        rawWeightA:
          type: number
          format: double
        rawWeightB:
          type: number
          format: double
        normalizedWeightA:
          type: number
          format: double
        normalizedWeightB:
          type: number
          format: double
    TfIdfCosineTrace:
      type: object
      required: [algorithmId, corpusSize, terms, dotProduct, rawNormA, rawNormB, cosine, angleDegrees, stemming]
      properties:
        algorithmId:
          $ref: "#/components/schemas/AlgorithmId"
        stemming:
          type: boolean
          description: >-
            Whether Porter stemming was applied to the token streams behind
            this result (the server's fixed `legajo.preprocess.stemming`
            setting; `false` by default). Embedding capabilities do not
            preprocess text, so for them it only echoes the setting.
        corpusSize:
          type: integer
        terms:
          type: array
          description: Only terms present in at least one of the two compared documents.
          items:
            $ref: "#/components/schemas/TfIdfTermTrace"
        dotProduct:
          type: number
          format: double
        rawNormA:
          type: number
          format: double
        rawNormB:
          type: number
          format: double
        cosine:
          type: number
          format: double
        angleDegrees:
          type: number
          format: double
    EmbeddingLocalTrace:
      type: object
      required:
        - algorithmId
        - stemming
        - provider
        - model
        - dimension
        - vectorAExcerpt
        - vectorBExcerpt
        - vectorA
        - vectorB
        - preNormL2A
        - preNormL2B
        - dotProduct
        - cosine
        - angleDegrees
        - normalizedScore
      properties:
        algorithmId:
          $ref: "#/components/schemas/AlgorithmId"
        stemming:
          type: boolean
          description: >-
            Whether Porter stemming was applied to the token streams behind
            this result (the server's fixed `legajo.preprocess.stemming`
            setting; `false` by default). Embedding capabilities do not
            preprocess text, so for them it only echoes the setting.
        provider:
          type: string
        model:
          type: string
        dimension:
          type: integer
        vectorAExcerpt:
          type: array
          items:
            type: number
            format: double
        vectorBExcerpt:
          type: array
          items:
            type: number
            format: double
        vectorA:
          type: array
          items:
            type: number
            format: double
        vectorB:
          type: array
          items:
            type: number
            format: double
        preNormL2A:
          type: number
          format: double
        preNormL2B:
          type: number
          format: double
        dotProduct:
          type: number
          format: double
        cosine:
          type: number
          format: double
        angleDegrees:
          type: number
          format: double
        normalizedScore:
          type: number
          format: double
    EmbeddingApiTrace:
      type: object
      required:
        - algorithmId
        - stemming
        - provider
        - model
        - dimension
        - vectorAExcerpt
        - vectorBExcerpt
        - vectorA
        - vectorB
        - preNormL2A
        - preNormL2B
        - sumSquaredDiff
        - distance
        - normalizedScore
        - providerStatus
      properties:
        algorithmId:
          $ref: "#/components/schemas/AlgorithmId"
        stemming:
          type: boolean
          description: >-
            Whether Porter stemming was applied to the token streams behind
            this result (the server's fixed `legajo.preprocess.stemming`
            setting; `false` by default). Embedding capabilities do not
            preprocess text, so for them it only echoes the setting.
        provider:
          type: string
        model:
          type: string
        dimension:
          type: integer
        vectorAExcerpt:
          type: array
          items:
            type: number
            format: double
        vectorBExcerpt:
          type: array
          items:
            type: number
            format: double
        vectorA:
          type: array
          items:
            type: number
            format: double
        vectorB:
          type: array
          items:
            type: number
            format: double
        preNormL2A:
          type: number
          format: double
        preNormL2B:
          type: number
          format: double
        sumSquaredDiff:
          type: number
          format: double
        distance:
          type: number
          format: double
        normalizedScore:
          type: number
          format: double
        providerStatus:
          type: string
    AlgorithmTrace:
      description: >-
        One variant per similarity capability.
        `levenshtein` and `needleman-wunsch` share the `DpMatrixTrace` shape.
      oneOf:
        - $ref: "#/components/schemas/DpMatrixTrace"
        - $ref: "#/components/schemas/JaccardTrace"
        - $ref: "#/components/schemas/TfIdfCosineTrace"
        - $ref: "#/components/schemas/EmbeddingLocalTrace"
        - $ref: "#/components/schemas/EmbeddingApiTrace"
      discriminator:
        propertyName: algorithmId
        mapping:
          levenshtein: "#/components/schemas/DpMatrixTrace"
          needleman-wunsch: "#/components/schemas/DpMatrixTrace"
          jaccard: "#/components/schemas/JaccardTrace"
          tfidf-cosine: "#/components/schemas/TfIdfCosineTrace"
          embedding-local: "#/components/schemas/EmbeddingLocalTrace"
          embedding-api: "#/components/schemas/EmbeddingApiTrace"
...